Separated react components into different files.
